Dev in disguise into main (#17)
* Testing for repo

* Get active users - repo

* Get active users from controller

* LandingPage only pulls active users for filter now.

* Runs

* ToDTO entity methods.

* Project repo beautification

* User repo beautification.

* Fix failing post project test by adding claim principal for test user

* Remove unused imports

* Fix comments

* Clean up some more

* Updated script to use universal URI

* Updated readme

* WithUserClaims() now creates an unauthorized user

* Cleaned up. Streamlined namespace notation.

* Moved to new folder to streamline structure

* Created new migrations

* Set properties to nullable, since they are already checked by database

* Implemented roles enum for users.

* Implemented roles for users to filter by

* Implemented read by email

* Now implements IComparable because Tag is used in sorted list

* Implemented role for users and set email index

* Updated migrations to support email and roles

* Cleaned up and added tests

* Added unknown role for error

* Added new test data for student role

* Fixed 400 bug, when creating Project with email

* Now support filtering by multiple roles

* Fixed bug with ordering of users causing 500 error

* Minor updates to script

* Made projects and taggroups associate itself to a university on creation

* things

* Delete shared folder

* revert tests

* Manage tag groups add/delete tags fixed

* Added license (#14)

* Fixed loading loop. Now displays relevant messages

* Fixed equals throws exception (TY Viktor <3)

* Cascade deletion of Tags now works

* Added todo

* Update button wrap and edit icon to make edit name more obvious

* Implement Equals method in DTOs
Implement Include in TagGroup readAsync Id
Implement Integration TEsts taggroup

* Change from taggroupdto to taggroupcreatedto

* Add claims to TestAuthHandler.cs to fix test issues
Remove TagGroupId from TagDTO in TagDTO.cs
Finish(?) TagGroupTests.cs

* Added equals methods

* Cleaned up

* Implemented read all by university. Tested.

* Implemented to dto for university

* Bug fixes, clean up and test writing

* Fixed bug where Supervisors where not loaded in repository

* Created fake university for Peanutbutter

* Added fake data for Peanutbutter

* Added SuperAdmin claim in integration test

* Fix integration test to run with sequential tag list

* Remove all/contains, replace with sequence

* Fixed old test results

* Fix no tag group name and http errors on manage tag groups

* Fix authorization view on projects overview in navbar

* Update build-and-test.yml

* fix ProjectControllerTests.cs to include email in dtos
change Equals method in ProjectDTO.cs to use Count and Contains instead of SetEquals

* create branch integration-testing
re-implement Includes in Read methods in TagGroupRepository.cs and ProjectRepository.cs

* Updated run script for final submission

* Finish integration tests
Add dummy users to factory

* Minor changes to user integration testing

* Update README.md

Add "about this project" section for short description of project scope/code, and tp add further development section/planned implementations/not yet implemented

* concurrency better now

* plus exstension

* "Something went wrong, please try reloading!"

* Fix ControllerTests

* Updated readme, needs readthrough

* optimised statements recommended by rider

* Tag Group Repo beautification and comments.

* controller constructors changed to lambda.

* updated read me and removed code

* Update README.md

* fixed tests again

* Beautified User path somewhat.

* Implemented lambda functions where possible and removed unnecessary if braces.

* Removed unnecessary comments

* Remove unnecessary imports and a random comment

* Max length on project names and descriptions.

* Streamline DTOs and Entities field with StringLength(100)

* fix userController to return empty list

* Updated script

* Updated readme

* All tests green

+++ run.ps1
@@ -1,22 +1,28 @@
-param([bool]$production, [bool]$dev)
+# Run with Docker-compose in production mode: -docker $true
+# Run with Docker-compose in development mode: -docker $true -dev $true
+
+# Run with Dotnet run: <No arguments>
+# Run with Dotnet run with hot reload: -hotreload $true
+
+param([bool]$docker, [bool]$dev, [bool]$hotreload)
 
 function Main {
-    if (!$dev) {
-        Write-Host "STARTING VIA DOCKER-COMPOSE"
-        DockerCompose
+    if ($docker) {
+        Write-Host "STARTING VIA DOCKER-COMPOSE UP"
+        RunWithDockerCompose
     } else {
-        Write-Host "STARTING VIA DOTNET RUN WITH HOT RELOAD"
-        DotnetRun
+        Write-Host "STARTING VIA DOTNET RUN"
+        RunWithDotnetRun
     }
 }
 
-function DockerCompose {
-    Write-Host "PRODUCTION MODE = $production"
+function RunWithDockerCompose {
+    Write-Host "PRODUCTION MODE =" (-not $dev)
     Write-Host
-    $file = "docker-compose.dev.yml"
-
-    if ($production) {
-        $file = "docker-compose.prod.yml"
+    $file = "docker-compose.prod.yml"
+    
+    if ($dev) {
+        $file = "docker-compose.dev.yml"
     }
 
     docker-compose -f $file up -d --build
@@ -31,21 +37,31 @@ function DockerCompose {
     Write-Host
 }
 
-function DotnetRun {
-    $Env:ASPNETCORE_ENVIRONMENT = "Development"
-
+function RunWithDotnetRun {
+    if ($dev) {
+        Write-Host "HOT RELOAD = $hotreload"
+    }
+    Write-Host
+    
+    $project = "ProjectBank.Server"
     $user = "postgres"
     $password = New-Guid
     $db = "projectbank"
 
     Write-Host "STARTING DATABASE"
-    docker run --name db --rm -d -p 5431:5432 -e "POSTGRES_USER=$user" -e "POSTGRES_PASSWORD=$password" -e "POSTGRES_DB=$db" postgres:latest
+    if ($dev) {
+        # Delete container when stopped
+        docker run --name db --rm -d -p 5431:5432 -e "POSTGRES_USER=$user" -e "POSTGRES_PASSWORD=$password" -e "POSTGRES_DB=$db" postgres:latest
+    } else {
+        docker run --name db -d -p 5431:5432 -e "POSTGRES_USER=$user" -e "POSTGRES_PASSWORD=$password" -e "POSTGRES_DB=$db" postgres:latest
+    }
+    
     Write-Host "DONE."
     Write-Host 
 
-    Write-Host "SETTING DOTNET SECRETS"
+    Write-Host "CONFIGURING CONNECTION STRING"
     $connectionString = "Host=localhost;Port=5431;Database=$db;Username=$user;Password=$password"
-    dotnet user-secrets init --project ProjectBank.Server
+    dotnet user-secrets init --project $project
     dotnet user-secrets set "ConnectionStrings:ProjectBank" "$connectionString" --project ProjectBank.Server
     Write-Host "DONE."
 
@@ -57,8 +73,13 @@ function DotnetRun {
 
     Write-Host "STARTING APPLICATION"
     
-    dotnet watch run --project ./ProjectBank.Server/
-
+    if ($dev) {
+        $Env:ASPNETCORE_ENVIRONMENT = "Development"
+        dotnet watch run --project $project
+    } else {
+        dotnet run --project $project
+    }
+    
     Write-Host
     Write-Host "STOPPING AND DELETING DATABASE"
     docker stop db
